code/bgpdownloader-master/requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
minio==7.1.13
mysql-connector==2.2.9
PyMySQL==1.0.2
python-dateutil==2.8.2
requests==2.28.2
urllib3==1.26.14
tqdm==4.66.4